\name{RCopyFind-package}
\alias{RCopyFind-package}
\alias{RCopyFind}
\docType{package}
\title{
\packageTitle{RCopyFind}
}
\description{
\packageDescription{RCopyFind}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCopyFind}
\packageIndices{RCopyFind}
}
\author{
\packageAuthor{RCopyFind}

Maintainer: \packageMaintainer{RCopyFind}
}
\references{
<<<<<<< HEAD
DeMora, Stephanie L., Loren Collingwood, and Adriana Ninci. "Shoot First in ALEC's Castle: The
=======
Collingwood, Loren, Stephanie DeMora, and Adriana Ninci. "Shoot First in ALECs Castle: The
>>>>>>> b893dfc2dbe520857ee8a32b83e45df6bc92eab4
Diffusion of Stand Your Ground Laws."
}
\keyword{ package }
\seealso{
}
\examples{
}
