\name{RCopyFind-package}
\alias{RCopyFind-package}
\alias{RCopyFind}
\docType{package}
\title{
\packageTitle{RCopyFind}
}
\description{
\packageDescription{RCopyFind}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCopyFind}
\packageIndices{RCopyFind}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RCopyFind}

Maintainer: \packageMaintainer{RCopyFind}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
