\name{RCopyFind-package}
\alias{RCopyFind-package}
\alias{RCopyFind}
\docType{package}
\title{
\packageTitle{RCopyFind}
}
\description{
\packageDescription{RCopyFind}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCopyFind}
\packageIndices{RCopyFind}
}
\author{
\packageAuthor{RCopyFind}
Maintainer: \packageMaintainer{RCopyFind}
}
\references{
DeMora, Stephanie L., Loren Collingwood, and Adriana Ninci. "Shoot First in ALEC's Castle: The Diffusion of Stand Your Ground Laws."
}
\keyword{ package }
\seealso{
}
\examples{
}
